requirements.txt
fastapi
uvicorn
bcrypt
pydantic
numpy
tensorflow
pillow
pandas
joblib
geopy
textblob
requests
google-cloud-firestore
